requirements.txt
requests==2.31.0
praw==7.7.1
pandas==2.1.0
numpy==1.24.3
textblob==0.17.1
nltk==3.8.1
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.0
dash-bootstrap-components==1.5.0
flask==2.3.3
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.0
vaderSentiment==3.3.2 